move main interval subscription processing to controller
